+++ packages/cli/uni-builder/src/rspack/plugins/babel-post.ts
@@ -12,16 +12,17 @@ export const pluginBabelPost = (): RsbuildPlugin => ({
     api.modifyBundlerChain({
       handler: async (chain, { CHAIN_ID }) => {
         if (chain.module.rules.get(CHAIN_ID.RULE.JS)) {
-          const babelLoaderOptions = chain.module
+          const { cacheIdentifier, ...babelLoaderOptions } = chain.module
             .rule(CHAIN_ID.RULE.JS)
             .use(CHAIN_ID.USE.BABEL)
             .get('options');
+
           const config = api.getNormalizedConfig();
 
           if (
             babelLoaderOptions &&
             lodash.isEqual(
-              getDefaultBabelOptions(config.source.decorators),
+              getDefaultBabelOptions(config, api.context),
               babelLoaderOptions,
             )
           ) {