+++ packages/cli/uni-builder/src/webpack/index.ts
@@ -106,6 +106,7 @@ export async function createWebpackBuilder(
   const { default: HtmlWebpackPlugin } = await import('html-webpack-plugin');
 
   // Some third-party plug-ins depend on html-webpack-plugin, like sri
+  // @ts-expect-error compilation type mismatch
   setHTMLPlugin(HtmlWebpackPlugin);
 
   rsbuildConfig.provider = webpackProvider;