+++ packages/cli/uni-builder/tests/postcssLegacy.test.ts
@@ -19,4 +19,24 @@ describe('plugin-postcssLegacy', () => {
 
     expect(bundlerConfigs[0].module).toMatchSnapshot();
   });
+
+  it('should allow tools.postcss to override the plugins', async () => {
+    const rsbuild = await createUniBuilder({
+      bundlerType: 'rspack',
+      config: {
+        tools: {
+          postcss(config) {
+            config.postcssOptions!.plugins = [];
+          },
+        },
+      },
+      cwd: '',
+    });
+
+    const {
+      origin: { bundlerConfigs },
+    } = await rsbuild.inspectConfig();
+
+    expect(bundlerConfigs[0].module).toMatchSnapshot();
+  });
 });