+++ packages/generator/generators/monorepo-generator/src/index.ts
@@ -19,17 +19,9 @@ import {
 import {
   getPackageManagerText,
   getModernVersion,
+  getGeneratorPath,
 } from '@modern-js/generator-utils';
 import { i18n, localeKeys } from './locale';
-
-const getGeneratorPath = (generator: string, distTag: string) => {
-  if (process.env.CODESMITH_ENV === 'development') {
-    return path.dirname(require.resolve(generator));
-  } else if (distTag) {
-    return `${generator}@${distTag}`;
-  }
-  return generator;
-};
 
 export const handleTemplateFile = async (
   context: GeneratorContext,
@@ -75,7 +67,7 @@ export const handleTemplateFile = async (
 
   const { packageManager } = ans;
   await appApi.runSubGenerator(
-    getGeneratorPath(BaseGenerator, context.config.distTag),
+    getGeneratorPath(BaseGenerator, context.config.distTag, [__dirname]),
     undefined,
     { ...context.config, hasPlugin: false },
   );
@@ -129,13 +121,13 @@ export const handleTemplateFile = async (
   }
 
   await appApi.runSubGenerator(
-    getGeneratorPath(ChangesetGenerator, context.config.distTag),
+    getGeneratorPath(ChangesetGenerator, context.config.distTag, [__dirname]),
   );
 
   const { packagesInfo } = context.config;
   if (packagesInfo && Object.keys(packagesInfo).length > 0) {
     await appApi.runSubGenerator(
-      getGeneratorPath(PackagesGenerator, context.config.distTag),
+      getGeneratorPath(PackagesGenerator, context.config.distTag, [__dirname]),
       undefined,
       context.config,
     );