+++ packages/generator/generators/mwa-generator/src/index.ts
@@ -19,17 +19,9 @@ import {
   validatePackagePath,
   getPackageManagerText,
   getModernVersion,
+  getGeneratorPath,
 } from '@modern-js/generator-utils';
 import { i18n, localeKeys } from './locale';
-
-const getGeneratorPath = (generator: string, distTag: string) => {
-  if (process.env.CODESMITH_ENV === 'development') {
-    return path.dirname(require.resolve(generator));
-  } else if (distTag) {
-    return `${generator}@${distTag}`;
-  }
-  return generator;
-};
 
 export const handleTemplateFile = async (
   context: GeneratorContext,
@@ -131,7 +123,7 @@ export const handleTemplateFile = async (
   const dirname = path.basename(generator.outputPath);
 
   await appApi.runSubGenerator(
-    getGeneratorPath(BaseGenerator, context.config.distTag),
+    getGeneratorPath(BaseGenerator, context.config.distTag, [__dirname]),
     undefined,
     { ...context.config, hasPlugin: false },
   );
@@ -179,7 +171,7 @@ export const handleTemplateFile = async (
   }
 
   await appApi.runSubGenerator(
-    getGeneratorPath(EntryGenerator, context.config.distTag),
+    getGeneratorPath(EntryGenerator, context.config.distTag, [__dirname]),
     `./${projectPath}`,
     {
       ...context.config,
@@ -196,7 +188,7 @@ export const handleTemplateFile = async (
 
   if (packagesInfo && Object.keys(packagesInfo).length > 0) {
     await appApi.runSubGenerator(
-      getGeneratorPath(PackagesGenerator, context.config.distTag),
+      getGeneratorPath(PackagesGenerator, context.config.distTag, [__dirname]),
       undefined,
       context.config,
     );