+++ packages/solutions/monorepo-tools/bin/modern.js
@@ -1,5 +1,11 @@
 #!/usr/bin/env node
 
+const { version } = require('../package.json');
+
+if (!process.env.MODERN_JS_VERSION) {
+  process.env.MODERN_JS_VERSION = version;
+}
+
 require('@modern-js/core/runBin').run({
-  initialLog: `Modern.js Monorepo v${require('../package.json').version}`,
+  initialLog: `Modern.js Monorepo v${version}`,
 });