+++ packages/cli/plugin-data-loader/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @modern-js/plugin-data-loader
 
+## 2.48.2
+
+### Patch Changes
+
+- @modern-js/runtime-utils@2.48.2
+- @modern-js/utils@2.48.2
+
+## 2.48.1
+
+### Patch Changes
+
+- Updated dependencies [8942b90]
+- Updated dependencies [ce426f7]
+- Updated dependencies [74749ae]
+  - @modern-js/utils@2.48.1
+  - @modern-js/runtime-utils@2.48.1
+
+## 2.48.0
+
+### Patch Changes
+
+- Updated dependencies [c323a23]
+  - @modern-js/utils@2.48.0
+  - @modern-js/runtime-utils@2.48.0
+
+## 2.47.1
+
+### Patch Changes
+
+- @modern-js/runtime-utils@2.47.1
+- @modern-js/utils@2.47.1
+
 ## 2.47.0
 
 ### Patch Changes