+++ packages/generator/generator-cases/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @modern-js/generator-cases
 
+## 3.3.13
+
+### Patch Changes
+
+- @modern-js/generator-common@3.3.13
+
+## 3.3.12
+
+### Patch Changes
+
+- Updated dependencies [a873989]
+  - @modern-js/generator-common@3.3.12
+
+## 3.3.11
+
+### Patch Changes
+
+- @modern-js/generator-common@3.3.11
+
+## 3.3.10
+
+### Patch Changes
+
+- @modern-js/generator-common@3.3.10
+
 ## 3.3.9
 
 ### Patch Changes