+++ packages/generator/generator-common/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @modern-js/generator-common
 
+## 3.3.13
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.48.2
+
+## 3.3.12
+
+### Patch Changes
+
+- a873989: chore(generator): remove legacy storybook generator
+
+  chore(generator): 移除老的 storybook v6 生成器
+
+  - @modern-js/plugin-i18n@2.48.1
+
+## 3.3.11
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.48.0
+
+## 3.3.10
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.47.1
+
 ## 3.3.9
 
 ### Patch Changes