+++ packages/generator/generator-utils/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @modern-js/generator-utils
 
+## 3.3.13
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.48.2
+- @modern-js/utils@2.48.2
+- @modern-js/generator-common@3.3.13
+
+## 3.3.12
+
+### Patch Changes
+
+- Updated dependencies [8942b90]
+- Updated dependencies [ce426f7]
+- Updated dependencies [a873989]
+  - @modern-js/utils@2.48.1
+  - @modern-js/generator-common@3.3.12
+  - @modern-js/plugin-i18n@2.48.1
+
+## 3.3.11
+
+### Patch Changes
+
+- c246812: chore(generator): extract getGeneratorPath function to generator-utils
+  fechoreat(generator): 提取 getGeneratorPath 方法到 generator-utils 公共包
+- Updated dependencies [c323a23]
+  - @modern-js/utils@2.48.0
+  - @modern-js/plugin-i18n@2.48.0
+  - @modern-js/generator-common@3.3.11
+
+## 3.3.10
+
+### Patch Changes
+
+- @modern-js/generator-common@3.3.10
+- @modern-js/plugin-i18n@2.47.1
+- @modern-js/utils@2.47.1
+
 ## 3.3.9
 
 ### Patch Changes