+++ packages/review/eslint-config-app/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js-app/eslint-config
 
+## 2.48.2
+
+## 2.48.1
+
+## 2.48.0
+
+## 2.47.1
+
 ## 2.47.0
 
 ### Patch Changes