+++ packages/runtime/plugin-testing/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @modern-js/plugin-testing
 
+## 2.48.2
+
+### Patch Changes
+
+- Updated dependencies [8eccb3b]
+- Updated dependencies [08e062f]
+  - @modern-js/runtime@2.48.2
+  - @modern-js/bff-core@2.48.2
+  - @modern-js/prod-server@2.48.2
+  - @modern-js/babel-compiler@2.48.2
+  - @modern-js/plugin@2.48.2
+  - @modern-js/types@2.48.2
+  - @modern-js/utils@2.48.2
+
+## 2.48.1
+
+### Patch Changes
+
+- Updated dependencies [8942b90]
+- Updated dependencies [24b2ec1]
+- Updated dependencies [488ec21]
+- Updated dependencies [ce426f7]
+- Updated dependencies [74749ae]
+  - @modern-js/utils@2.48.1
+  - @modern-js/prod-server@2.48.1
+  - @modern-js/bff-core@2.48.1
+  - @modern-js/runtime@2.48.1
+  - @modern-js/babel-compiler@2.48.1
+  - @modern-js/plugin@2.48.1
+  - @modern-js/types@2.48.1
+
+## 2.48.0
+
+### Patch Changes
+
+- Updated dependencies [0b44ddb]
+- Updated dependencies [c323a23]
+  - @modern-js/prod-server@2.48.0
+  - @modern-js/utils@2.48.0
+  - @modern-js/runtime@2.48.0
+  - @modern-js/bff-core@2.48.0
+  - @modern-js/babel-compiler@2.48.0
+  - @modern-js/plugin@2.48.0
+  - @modern-js/types@2.48.0
+
+## 2.47.1
+
+### Patch Changes
+
+- Updated dependencies [cee9184]
+  - @modern-js/runtime@2.47.1
+  - @modern-js/bff-core@2.47.1
+  - @modern-js/prod-server@2.47.1
+  - @modern-js/babel-compiler@2.47.1
+  - @modern-js/plugin@2.47.1
+  - @modern-js/types@2.47.1
+  - @modern-js/utils@2.47.1
+
 ## 2.47.0
 
 ### Patch Changes