+++ packages/solutions/app-tools/CHANGELOG.md
@@ -1,5 +1,101 @@
 # @modern-js/app-tools
 
+## 2.48.2
+
+### Patch Changes
+
+- Updated dependencies [9e7115f]
+  - @modern-js/uni-builder@2.48.2
+  - @modern-js/server@2.48.2
+  - @modern-js/core@2.48.2
+  - @modern-js/plugin-data-loader@2.48.2
+  - @modern-js/plugin-i18n@2.48.2
+  - @modern-js/plugin-lint@2.48.2
+  - @modern-js/server-core@2.48.2
+  - @modern-js/prod-server@2.48.2
+  - @modern-js/server-utils@2.48.2
+  - @modern-js/node-bundle-require@2.48.2
+  - @modern-js/plugin@2.48.2
+  - @modern-js/types@2.48.2
+  - @modern-js/utils@2.48.2
+
+## 2.48.1
+
+### Patch Changes
+
+- 9835dbf: feat(uni-builder): bump rsbuild 0.4.11 and support modify rspack builtin:swc-loadder config by tools.swc
+
+  feat(uni-builder): rsbuild 版本升级到 0.4.11，并支持通过 tools.swc 修改 rspack builtin:swc-loadder 配置
+
+- c86735b: fix: adpater modern worker, worker bundle need export serverRender fn
+  fix: 适配 modern worker, worker bundle 需要导出 serverRender 函数
+- Updated dependencies [8942b90]
+- Updated dependencies [24b2ec1]
+- Updated dependencies [ce426f7]
+- Updated dependencies [9835dbf]
+- Updated dependencies [77209fe]
+- Updated dependencies [fc93013]
+- Updated dependencies [74749ae]
+- Updated dependencies [788b3be]
+- Updated dependencies [0d0a886]
+- Updated dependencies [02060c5]
+  - @modern-js/utils@2.48.1
+  - @modern-js/prod-server@2.48.1
+  - @modern-js/server@2.48.1
+  - @modern-js/uni-builder@2.48.1
+  - @modern-js/core@2.48.1
+  - @modern-js/plugin-data-loader@2.48.1
+  - @modern-js/plugin-i18n@2.48.1
+  - @modern-js/plugin-lint@2.48.1
+  - @modern-js/server-core@2.48.1
+  - @modern-js/server-utils@2.48.1
+  - @modern-js/node-bundle-require@2.48.1
+  - @modern-js/plugin@2.48.1
+  - @modern-js/types@2.48.1
+
+## 2.48.0
+
+### Patch Changes
+
+- c323a23: feat: separate new & upgrade action from solutions
+  feat: 从工程方案中分离 new & upgrade 命令
+- Updated dependencies [ecaf916]
+- Updated dependencies [0b44ddb]
+- Updated dependencies [c323a23]
+  - @modern-js/uni-builder@2.48.0
+  - @modern-js/prod-server@2.48.0
+  - @modern-js/utils@2.48.0
+  - @modern-js/core@2.48.0
+  - @modern-js/server@2.48.0
+  - @modern-js/plugin-data-loader@2.48.0
+  - @modern-js/plugin-i18n@2.48.0
+  - @modern-js/plugin-lint@2.48.0
+  - @modern-js/server-core@2.48.0
+  - @modern-js/server-utils@2.48.0
+  - @modern-js/node-bundle-require@2.48.0
+  - @modern-js/plugin@2.48.0
+  - @modern-js/types@2.48.0
+
+## 2.47.1
+
+### Patch Changes
+
+- @modern-js/upgrade@2.47.1
+- @modern-js/core@2.47.1
+- @modern-js/plugin-data-loader@2.47.1
+- @modern-js/plugin-i18n@2.47.1
+- @modern-js/plugin-lint@2.47.1
+- @modern-js/uni-builder@2.47.1
+- @modern-js/new-action@2.47.1
+- @modern-js/server-core@2.47.1
+- @modern-js/prod-server@2.47.1
+- @modern-js/server@2.47.1
+- @modern-js/server-utils@2.47.1
+- @modern-js/node-bundle-require@2.47.1
+- @modern-js/plugin@2.47.1
+- @modern-js/types@2.47.1
+- @modern-js/utils@2.47.1
+
 ## 2.47.0
 
 ### Patch Changes