+++ packages/solutions/module-tools/CHANGELOG.md
@@ -1,5 +1,67 @@
 # @modern-js/module-tools
 
+## 2.48.2
+
+### Patch Changes
+
+- @modern-js/core@2.48.2
+- @modern-js/plugin-changeset@2.48.2
+- @modern-js/plugin-i18n@2.48.2
+- @modern-js/plugin-lint@2.48.2
+- @modern-js/plugin@2.48.2
+- @modern-js/types@2.48.2
+- @modern-js/utils@2.48.2
+
+## 2.48.1
+
+### Patch Changes
+
+- b46c84e: fix(module-tools): support tsconfig extends.
+  fix(module-tools): 支持解析 tsconfig extends.
+- 1129519: fix(module-tools): target es5 not work for mjs chunk
+- 4833040: fix(module-tools): new matcher to match alias when exist comment in dynamic import
+  fix(module-tools): 新的匹配器以匹配别名即使动态引用里存在注释
+- Updated dependencies [8942b90]
+- Updated dependencies [ce426f7]
+  - @modern-js/utils@2.48.1
+  - @modern-js/core@2.48.1
+  - @modern-js/plugin-changeset@2.48.1
+  - @modern-js/plugin-i18n@2.48.1
+  - @modern-js/plugin-lint@2.48.1
+  - @modern-js/plugin@2.48.1
+  - @modern-js/types@2.48.1
+
+## 2.48.0
+
+### Patch Changes
+
+- d109126: feat(module-tools): add support for custom asset name
+  feat(module-tools): 支持自定义资源文件名
+- c323a23: feat: separate new & upgrade action from solutions
+  feat: 从工程方案中分离 new & upgrade 命令
+- Updated dependencies [c323a23]
+  - @modern-js/utils@2.48.0
+  - @modern-js/core@2.48.0
+  - @modern-js/plugin-changeset@2.48.0
+  - @modern-js/plugin-i18n@2.48.0
+  - @modern-js/plugin-lint@2.48.0
+  - @modern-js/plugin@2.48.0
+  - @modern-js/types@2.48.0
+
+## 2.47.1
+
+### Patch Changes
+
+- @modern-js/upgrade@2.47.1
+- @modern-js/core@2.47.1
+- @modern-js/plugin-changeset@2.47.1
+- @modern-js/plugin-i18n@2.47.1
+- @modern-js/plugin-lint@2.47.1
+- @modern-js/new-action@2.47.1
+- @modern-js/plugin@2.47.1
+- @modern-js/types@2.47.1
+- @modern-js/utils@2.47.1
+
 ## 2.47.0
 
 ### Patch Changes