+++ packages/toolkit/plugin/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @modern-js/plugin
 
+## 2.48.2
+
+### Patch Changes
+
+- @modern-js/utils@2.48.2
+
+## 2.48.1
+
+### Patch Changes
+
+- Updated dependencies [8942b90]
+- Updated dependencies [ce426f7]
+  - @modern-js/utils@2.48.1
+
+## 2.48.0
+
+### Patch Changes
+
+- Updated dependencies [c323a23]
+  - @modern-js/utils@2.48.0
+
+## 2.47.1
+
+### Patch Changes
+
+- @modern-js/utils@2.47.1
+
 ## 2.47.0
 
 ### Patch Changes