+++ packages/toolkit/runtime-utils/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @modern-js/runtime-utils
 
+## 2.48.2
+
+### Patch Changes
+
+- @modern-js/utils@2.48.2
+
+## 2.48.1
+
+### Patch Changes
+
+- 74749ae: fix: sometime the url '/a/' shoud be equal '/a'
+  fix: 有时 url '/a/' 应该等于 '/a'
+- Updated dependencies [8942b90]
+- Updated dependencies [ce426f7]
+  - @modern-js/utils@2.48.1
+
+## 2.48.0
+
+### Patch Changes
+
+- Updated dependencies [c323a23]
+  - @modern-js/utils@2.48.0
+
+## 2.47.1
+
+### Patch Changes
+
+- @modern-js/utils@2.47.1
+
 ## 2.47.0
 
 ### Patch Changes